figured out how to add sql
the name for the medical corporation can't be updated in "Owners" first because it's a child, but then I couldn't add the ownership Id . So I got rid of the foreign keys in Owners, and made the ownership ID a foreign key in both medical_corporations and physician_owners. If I want to delete any of that, I will need to first delete from Owners.

I also figured out how to properly get the last_insert_id into the table via PHP. I will need to copy this properly into the other files.